test.sh: don't require coverage
Only run atomic under coverage if it's installed. This will allow the
testsuite to be run on Atomic Host.

Also allow passing the $PYTHON env var to specify which python
executable to use when running the test. It was implicitly using the
same Python as the one for coverage before, which was most always py2.