tests/storage: check for grep failure
Split the pipeline in two so that we check that there's something to
unmount before doing it. This is needed now because the new CI
infrastructure provisions a true uninitialized block device that doesn't
need to be unmounted, so the `grep` would fail (because of the
`pipefail` option). Previously, the `/dev/vdb` provided was already
pre-formatted and thus automounted.

This should catch both the old CI style, as well as local developer
setups with more conventional partitions occupying `/dev/vdb`.

I dropped the `-r` since we *know* at that point that there is at least
one entry to unmount.