tests/storage: check for grep failure
Split the pipeline in two so that we check that there's something to
unmount before doing it. This is needed now because the new CI
infrastructure provisions a true uninitialized block device that doesn't
need to be unmounted, so the `grep` would fail (because of the
`pipefail` option). Previously, `/dev/vdb` was actually itself a
pre-formatted ext4 partition and thus was automounted.

This should now catch both the old CI style (in case local developers
were mirroring that setup), as well as more conventional partitioning
schemes.

I dropped the `-r` since we *know* at that point that there is at least
one entry to unmount.